+++ sdk/identity/azure-identity/azure/identity/_credentials/azure_powershell.py
@@ -4,6 +4,7 @@
 # ------------------------------------
 import base64
 import logging
+import os
 import platform
 import subprocess
 import sys
@@ -18,6 +19,7 @@ from .azure_cli import get_safe_working_dir
 from .. import CredentialUnavailableError
 from .._internal import _scopes_to_resource, resolve_tenant
 from .._internal.decorators import log_get_token
+from .._constants import EnvironmentVariables
 
 if TYPE_CHECKING:
     # pylint:disable=ungrouped-imports
@@ -51,15 +53,12 @@ class AzurePowerShellCredential(object):
     """Authenticates by requesting a token from Azure PowerShell.
 
     This requires previously logging in to Azure via "Connect-AzAccount", and will use the currently logged in identity.
-
-    :keyword bool allow_multitenant_authentication: when True, enables the credential to acquire tokens from any tenant
-        the identity logged in to Azure PowerShell is registered in. When False, which is the default, the credential
-        will acquire tokens only from the tenant of Azure PowerShell's active subscription.
     """
 
     def __init__(self, **kwargs):
         # type: (**Any) -> None
-        self._allow_multitenant = kwargs.get("allow_multitenant_authentication", False)
+        disable_multitenant = os.environ.get(EnvironmentVariables.AZURE_IDENTITY_DISABLE_MULTITENANTAUTH, False)
+        self._allow_multitenant = not disable_multitenant
 
     def __enter__(self):
         return self
@@ -80,8 +79,7 @@ class AzurePowerShellCredential(object):
         also handle token caching because this credential doesn't cache the tokens it acquires.
 
         :param str scopes: desired scope for the access token. This credential allows only one scope per request.
-        :keyword str tenant_id: optional tenant to include in the token request. If **allow_multitenant_authentication**
-            is False, specifying a tenant with this argument may raise an exception.
+        :keyword str tenant_id: optional tenant to include in the token request.
 
         :rtype: :class:`azure.core.credentials.AccessToken`