+++ sdk/identity/azure-identity/tests/test_aad_client.py
@@ -304,11 +304,12 @@ def test_multitenant_cache():
     assert client_b.get_cached_access_token([scope]) is None
 
     # but C allows multitenant auth and should therefore return the token from tenant_a when appropriate
-    client_c = AadClient(tenant_id=tenant_c, allow_multitenant_authentication=True, **common_args)
+    client_c = AadClient(tenant_id=tenant_c, **common_args)
     assert client_c.get_cached_access_token([scope]) is None
     token = client_c.get_cached_access_token([scope], tenant_id=tenant_a)
     assert token.token == expected_token
     with patch.dict(
-        "os.environ", {EnvironmentVariables.AZURE_IDENTITY_ENABLE_LEGACY_TENANT_SELECTION: "true"}, clear=True
+        "os.environ", {EnvironmentVariables.AZURE_IDENTITY_DISABLE_MULTITENANTAUTH: "true"}, clear=True
     ):
-        assert client_c.get_cached_access_token([scope], tenant_id=tenant_a) is None
+        with pytest.raises(ClientAuthenticationError, match="multitenant_authentication"):
+            client_c.get_cached_access_token([scope], tenant_id=tenant_a)