+++ sdk/identity/azure-identity/tests/test_auth_code_async.py
@@ -143,8 +143,6 @@ async def test_auth_code_credential():
 
 
 async def test_allow_multitenant_authentication():
-    """When allow_multitenant_authentication is True, the credential should respect get_token(tenant_id=...)"""
-
     first_tenant = "first-tenant"
     first_token = "***"
     second_tenant = "second-tenant"
@@ -162,7 +160,6 @@ async def test_allow_multitenant_authentication():
         "client-id",
         "authcode",
         "https://localhost",
-        allow_multitenant_authentication=True,
         transport=Mock(send=send),
     )
     token = await credential.get_token("scope")
@@ -180,8 +177,6 @@ async def test_allow_multitenant_authentication():
 
 
 async def test_multitenant_authentication_not_allowed():
-    """get_token(tenant_id=...) should raise when allow_multitenant_authentication is False (the default)"""
-
     expected_tenant = "expected-tenant"
     expected_token = "***"
 
@@ -202,11 +197,10 @@ async def test_multitenant_authentication_not_allowed():
     token = await credential.get_token("scope", tenant_id=expected_tenant)
     assert token.token == expected_token
 
-    # but any other tenant should get an error
-    with pytest.raises(ClientAuthenticationError, match="allow_multitenant_authentication"):
-        await credential.get_token("scope", tenant_id="un" + expected_tenant)
+    token = await credential.get_token("scope", tenant_id="un" + expected_tenant)
+    assert token.token == expected_token * 2
 
-    # ...unless the compat switch is enabled
-    with patch.dict("os.environ", {EnvironmentVariables.AZURE_IDENTITY_ENABLE_LEGACY_TENANT_SELECTION: "true"}):
-        token = await credential.get_token("scope", tenant_id="un" + expected_tenant)
-    assert token.token == expected_token, "credential should ignore tenant_id kwarg when the compat switch is enabled"
+    # but fail if multi tenant is disabled
+    with patch.dict("os.environ", {EnvironmentVariables.AZURE_IDENTITY_DISABLE_MULTITENANTAUTH: "true"}):
+        with pytest.raises(ClientAuthenticationError, match="multitenant_authentication"):
+            await credential.get_token("scope", tenant_id="un" + expected_tenant)