+++ sdk/identity/azure-identity/tests/test_cli_credential.py
@@ -153,8 +153,6 @@ def test_timeout():
 
 
 def test_allow_multitenant_authentication():
-    """When allow_multitenant_authentication is True, the credential should respect get_token(tenant_id=...)"""
-
     default_tenant = "first-tenant"
     first_token = "***"
     second_tenant = "second-tenant"
@@ -174,7 +172,7 @@ def test_allow_multitenant_authentication():
             }
         )
 
-    credential = AzureCliCredential(allow_multitenant_authentication=True)
+    credential = AzureCliCredential()
     with mock.patch(CHECK_OUTPUT, fake_check_output):
         token = credential.get_token("scope")
         assert token.token == first_token
@@ -188,41 +186,3 @@ def test_allow_multitenant_authentication():
         # should still default to the first tenant
         token = credential.get_token("scope")
         assert token.token == first_token
-
-
-def test_multitenant_authentication_not_allowed():
-    """get_token(tenant_id=...) should raise when allow_multitenant_authentication is False (the default)"""
-
-    expected_tenant = "expected-tenant"
-    expected_token = "***"
-
-    def fake_check_output(command_line, **_):
-        match = re.search("--tenant (.*)", command_line[-1])
-        assert match is None or match[1] == expected_tenant
-        return json.dumps(
-            {
-                "expiresOn": datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f"),
-                "accessToken": expected_token,
-                "subscription": "some-guid",
-                "tenant": expected_token,
-                "tokenType": "Bearer",
-            }
-        )
-
-    credential = AzureCliCredential()
-    with mock.patch(CHECK_OUTPUT, fake_check_output):
-        token = credential.get_token("scope")
-        assert token.token == expected_token
-
-        # specifying a tenant should get an error
-        with pytest.raises(ClientAuthenticationError, match="allow_multitenant_authentication"):
-            credential.get_token("scope", tenant_id="un" + expected_tenant)
-
-        # ...unless the compat switch is enabled
-        with mock.patch.dict(
-            "os.environ", {EnvironmentVariables.AZURE_IDENTITY_ENABLE_LEGACY_TENANT_SELECTION: "true"}
-        ):
-            token = credential.get_token("scope", tenant_id="un" + expected_tenant)
-        assert (
-            token.token == expected_token
-        ), "credential should ignore tenant_id kwarg when the compat switch is enabled"