+++ sdk/identity/azure-identity/tests/test_client_secret_credential.py
@@ -4,7 +4,8 @@
 # ------------------------------------
 from azure.core.exceptions import ClientAuthenticationError
 from azure.core.pipeline.policies import ContentDecodePolicy, SansIOHTTPPolicy
-from azure.identity import ClientSecretCredential, RegionalAuthority, TokenCachePersistenceOptions
+from azure.identity import ClientSecretCredential, TokenCachePersistenceOptions
+from azure.identity._enums import RegionalAuthority
 from azure.identity._constants import EnvironmentVariables
 from azure.identity._internal.user_agent import USER_AGENT
 from msal import TokenCache
@@ -212,8 +213,6 @@ def test_cache_multiple_clients():
 
 
 def test_allow_multitenant_authentication():
-    """When allow_multitenant_authentication is True, the credential should respect get_token(tenant_id=...)"""
-
     first_tenant = "first-tenant"
     first_token = "***"
     second_tenant = "second-tenant"
@@ -230,7 +229,7 @@ def test_allow_multitenant_authentication():
         return mock_response(json_payload=build_aad_response(access_token=token))
 
     credential = ClientSecretCredential(
-        first_tenant, "client-id", "secret", allow_multitenant_authentication=True, transport=Mock(send=send)
+        first_tenant, "client-id", "secret", transport=Mock(send=send)
     )
     token = credential.get_token("scope")
     assert token.token == first_token
@@ -247,8 +246,6 @@ def test_allow_multitenant_authentication():
 
 
 def test_multitenant_authentication_not_allowed():
-    """get_token(tenant_id=...) should raise when allow_multitenant_authentication is False (the default)"""
-
     expected_tenant = "expected-tenant"
     expected_token = "***"
 
@@ -270,11 +267,9 @@ def test_multitenant_authentication_not_allowed():
     token = credential.get_token("scope", tenant_id=expected_tenant)
     assert token.token == expected_token
 
-    # but any other tenant should get an error
-    with pytest.raises(ClientAuthenticationError, match="allow_multitenant_authentication"):
-        credential.get_token("scope", tenant_id="un" + expected_tenant)
+    token = credential.get_token("scope", tenant_id="un" + expected_tenant)
+    assert token.token == expected_token
 
-    # ...unless the compat switch is enabled
-    with patch.dict("os.environ", {EnvironmentVariables.AZURE_IDENTITY_ENABLE_LEGACY_TENANT_SELECTION: "true"}):
-        token = credential.get_token("scope", tenant_id="un" + expected_tenant)
-    assert token.token == expected_token, "credential should ignore tenant_id kwarg when the compat switch is enabled"
+    with patch.dict("os.environ", {EnvironmentVariables.AZURE_IDENTITY_DISABLE_MULTITENANTAUTH: "true"}):
+        with pytest.raises(ClientAuthenticationError, match="multitenant_authentication"):
+            token = credential.get_token("scope", tenant_id="un" + expected_tenant)