+++ sdk/identity/azure-identity/tests/test_shared_cache_credential.py
@@ -826,8 +826,6 @@ def test_claims_challenge():
 
 
 def test_allow_multitenant_authentication():
-    """When allow_multitenant_authentication is True, the credential should respect get_token(tenant_id=...)"""
-
     default_tenant = "organizations"
     first_token = "***"
     second_tenant = "second-tenant"
@@ -851,7 +849,7 @@ def test_allow_multitenant_authentication():
     cache = populated_cache(expected_account)
 
     credential = SharedTokenCacheCredential(
-        allow_multitenant_authentication=True, authority=authority, transport=Mock(send=send), _cache=cache
+        authority=authority, transport=Mock(send=send), _cache=cache
     )
     token = credential.get_token("scope")
     assert token.token == first_token
@@ -867,56 +865,7 @@ def test_allow_multitenant_authentication():
     assert token.token == first_token
 
 
-def test_multitenant_authentication_not_allowed():
-    """get_token(tenant_id=...) should raise when allow_multitenant_authentication is False (the default)"""
-
-    default_tenant = "organizations"
-    expected_token = "***"
-
-    def send(request, **_):
-        parsed = urlparse(request.url)
-        tenant_id = parsed.path.split("/")[1]
-        assert tenant_id == default_tenant
-        return mock_response(
-            json_payload=build_aad_response(
-                access_token=expected_token,
-                id_token_claims=id_token_claims(aud="...", iss="...", sub="..."),
-            )
-        )
-
-    tenant_id = "tenant-id"
-    client_id = "client-id"
-    authority = "localhost"
-    object_id = "object-id"
-    username = "me"
-
-    expected_account = get_account_event(
-        username, object_id, tenant_id, authority=authority, client_id=client_id, refresh_token="**"
-    )
-    cache = populated_cache(expected_account)
-
-    credential = SharedTokenCacheCredential(authority=authority, transport=Mock(send=send), _cache=cache)
-
-    token = credential.get_token("scope")
-    assert token.token == expected_token
-
-    # explicitly specifying the configured tenant is okay
-    token = credential.get_token("scope", tenant_id=default_tenant)
-    assert token.token == expected_token
-
-    # but any other tenant should get an error
-    with pytest.raises(ClientAuthenticationError, match="allow_multitenant_authentication"):
-        credential.get_token("scope", tenant_id="some tenant")
-
-    # ...unless the compat switch is enabled
-    with patch.dict("os.environ", {EnvironmentVariables.AZURE_IDENTITY_ENABLE_LEGACY_TENANT_SELECTION: "true"}):
-        token = credential.get_token("scope", tenant_id="some tenant")
-    assert token.token == expected_token, "credential should ignore tenant_id kwarg when the compat switch is enabled"
-
-
 def test_allow_multitenant_authentication_auth_record():
-    """When allow_multitenant_authentication is True, the credential should respect get_token(tenant_id=...)"""
-
     default_tenant = "organizations"
     first_token = "***"
     second_tenant = "second-tenant"
@@ -947,7 +896,6 @@ def test_allow_multitenant_authentication_auth_record():
     cache = populated_cache(expected_account)
 
     credential = SharedTokenCacheCredential(
-        allow_multitenant_authentication=True,
         authority=authority,
         transport=Mock(send=send),
         authentication_record=record,
@@ -965,64 +913,6 @@ def test_allow_multitenant_authentication_auth_record():
     # should still default to the first tenant
     token = credential.get_token("scope")
     assert token.token == first_token
-
-
-def test_multitenant_authentication_not_allowed_authentication_record():
-    """get_token(tenant_id=...) should raise when allow_multitenant_authentication is False (the default)"""
-
-    default_tenant = "organizations"
-    expected_token = "***"
-
-    authority = AzureAuthorityHosts.AZURE_PUBLIC_CLOUD
-    object_id = "object-id"
-    home_account_id = object_id + "." + default_tenant
-    record = AuthenticationRecord(default_tenant, "client-id", authority, home_account_id, "user")
-
-    def send(request, **_):
-        parsed = urlparse(request.url)
-        tenant_id = parsed.path.split("/")[1]
-        if "/oauth2/v2.0/token" not in request.url:
-            return get_discovery_response("https://{}/{}".format(parsed.netloc, tenant_id))
-
-        assert tenant_id == default_tenant
-        return mock_response(
-            json_payload=build_aad_response(
-                access_token=expected_token,
-                id_token_claims=id_token_claims(aud="...", iss="...", sub="..."),
-            )
-        )
-
-    expected_account = get_account_event(
-        record.username,
-        object_id,
-        record.tenant_id,
-        authority=record.authority,
-        client_id=record.client_id,
-        refresh_token="**",
-    )
-    cache = populated_cache(expected_account)
-
-    credential = SharedTokenCacheCredential(
-        authority=authority, transport=Mock(send=send), authentication_record=record, _cache=cache
-    )
-
-    token = credential.get_token("scope")
-    assert token.token == expected_token
-
-    # explicitly specifying the configured tenant is okay
-    token = credential.get_token("scope", tenant_id=default_tenant)
-    assert token.token == expected_token
-
-    # but any other tenant should get an error
-    with pytest.raises(ClientAuthenticationError, match="allow_multitenant_authentication"):
-        credential.get_token("scope", tenant_id="some tenant")
-
-    # ...unless the compat switch is enabled
-    with patch.dict(
-        "os.environ", {EnvironmentVariables.AZURE_IDENTITY_ENABLE_LEGACY_TENANT_SELECTION: "true"}, clear=True
-    ):
-        token = credential.get_token("scope", tenant_id="some tenant")
-    assert token.token == expected_token, "credential should ignore tenant_id kwarg when the compat switch is enabled"
 
 
 def get_account_event(