+++ sdk/identity/azure-identity/tests/test_vscode_credential.py
@@ -278,8 +278,6 @@ def test_no_user_settings():
 
 
 def test_allow_multitenant_authentication():
-    """When allow_multitenant_authentication is True, the credential should respect get_token(tenant_id=...)"""
-
     first_tenant = "first-tenant"
     first_token = "***"
     second_tenant = "second-tenant"
@@ -293,7 +291,7 @@ def test_allow_multitenant_authentication():
         return mock_response(json_payload=build_aad_response(access_token=token))
 
     credential = get_credential(
-        tenant_id=first_tenant, allow_multitenant_authentication=True, transport=mock.Mock(send=send)
+        tenant_id=first_tenant, transport=mock.Mock(send=send)
     )
     with mock.patch(GET_REFRESH_TOKEN, lambda _: "**"):
         token = credential.get_token("scope")
@@ -312,8 +310,6 @@ def test_allow_multitenant_authentication():
 
 
 def test_multitenant_authentication_not_allowed():
-    """get_token(tenant_id=...) should raise when allow_multitenant_authentication is False (the default)"""
-
     expected_tenant = "expected-tenant"
     expected_token = "***"
 
@@ -333,11 +329,9 @@ def test_multitenant_authentication_not_allowed():
     token = credential.get_token("scope", tenant_id=expected_tenant)
     assert token.token == expected_token
 
-    # but any other tenant should get an error
-    with pytest.raises(ClientAuthenticationError, match="allow_multitenant_authentication"):
-        credential.get_token("scope", tenant_id="un" + expected_tenant)
+    token = credential.get_token("scope", tenant_id="un" + expected_tenant)
+    assert token.token == expected_token * 2
 
-    # ...unless the compat switch is enabled
-    with mock.patch.dict("os.environ", {EnvironmentVariables.AZURE_IDENTITY_ENABLE_LEGACY_TENANT_SELECTION: "true"}):
-        token = credential.get_token("scope", tenant_id="un" + expected_tenant)
-    assert token.token == expected_token, "credential should ignore tenant_id kwarg when the compat switch is enabled"
+    with mock.patch.dict("os.environ", {EnvironmentVariables.AZURE_IDENTITY_DISABLE_MULTITENANTAUTH: "true"}):
+        with pytest.raises(ClientAuthenticationError, match="multitenant_authentication"):
+            token = credential.get_token("scope", tenant_id="un" + expected_tenant)